Move logic for analyzing GroupBy expression to GroupByAnalyzer
